% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_draws.R
\name{add_draws}
\alias{add_draws}
\title{Add draws to a data frame in tidy format}
\usage{
add_draws(data, draws, value = ".value")
}
\arguments{
\item{data}{Data frame to add draws to, with M rows.}

\item{draws}{N by M matrix of draws, with M columns corresponding to the M rows in \code{data}, and N draws in each column.}

\item{value}{The name of the output column; default \code{".value"}.}
}
\value{
A data frame (actually, a \link[tibble:tibble]{tibble}) with a \code{.row} column (a
factor grouping rows from the input \code{data}), a \code{.draw} column (a unique index corresponding to each draw
from the distribution), and a column with its name specified by the \code{value} argument (default is \code{.value})
containing the values of draws from \code{draws}. The data frame is grouped by all rows in \code{data} plus the \code{.row} column.
}
\description{
Add draws from a matrix of draws (usually draws from a predictive distribution) to a data frame in tidy format. This is
a generic version of \code{\link[=add_predicted_draws]{add_predicted_draws()}} that can be used with model types
that have their own prediction functions that are not yet supported by tidybayes.
}
\details{
Given a data frame with M rows and an N by M matrix of N draws, adds a \code{.row}, \code{.draw}, and \code{.value}
column (or another name if \code{value} is set) to \code{data}, and expands \code{data} into a long-format dataframe of draws.

\code{add_epred_draws(df, m)} is roughly equivalent to \code{add_draws(df, posterior_epred(m, newdata = df))}, except
that \code{add_epred_draws} standardizes argument names and values across packages and has additional features for some
model types (like handling ordinal responses and distributional parameters in brms).

\code{add_predicted_draws(df, m)} is roughly equivalent to \code{add_draws(df, posterior_predict(m, newdata = df))}, except
that \code{add_predicted_draws} standardizes argument names and values across packages.
}
\seealso{
\code{\link[=add_predicted_draws]{add_predicted_draws()}}, \code{\link[=add_draws]{add_draws()}}
}
\author{
Matthew Kay
}
\keyword{"Greys")}
\keyword{#}
\keyword{%>%}
\keyword{*}
\keyword{+}
\keyword{.))}
\keyword{.value),}
\keyword{/}
\keyword{0.25)}
\keyword{1}
\keyword{1,}
\keyword{101))}
\keyword{500)}
\keyword{=}
\keyword{\dontrun{}
\keyword{across}
\keyword{add_draws(posterior_epred(m_mpg,}
\keyword{add_epred_draws(m_mpg),}
\keyword{alpha}
\keyword{arguments}
\keyword{below}
\keyword{brm(mpg}
\keyword{chain}
\keyword{chains}
\keyword{color}
\keyword{cyl,}
\keyword{data}
\keyword{data_grid(hp}
\keyword{do}
\keyword{does}
\keyword{equivalent}
\keyword{example}
\keyword{except}
\keyword{few}
\keyword{geom_point(data}
\keyword{ggplot(aes(x}
\keyword{group_by(cyl)}
\keyword{hp}
\keyword{hp,}
\keyword{in}
\keyword{intervals}
\keyword{is}
\keyword{it}
\keyword{iter}
\keyword{iterations}
\keyword{just}
\keyword{library(brms)}
\keyword{library(dplyr)}
\keyword{library(ggplot2)}
\keyword{library(modelr)}
\keyword{line}
\keyword{m_mpg}
\keyword{manip}
\keyword{model}
\keyword{mpg,}
\keyword{mtcars}
\keyword{mtcars)}
\keyword{mtcars,}
\keyword{n}
\keyword{newdata}
\keyword{not}
\keyword{ordered(cyl)))}
\keyword{plot}
\keyword{posterior}
\keyword{practice}
\keyword{predictive}
\keyword{quickly}
\keyword{roughly}
\keyword{runs}
\keyword{scale_fill_brewer(palette}
\keyword{seq_range(hp,}
\keyword{so}
\keyword{standardize}
\keyword{stat_lineribbon(aes(y}
\keyword{that}
\keyword{the}
\keyword{theme_set(theme_light())}
\keyword{to}
\keyword{types.}
\keyword{use}
\keyword{y}
\keyword{}}
\keyword{~}
