% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bins.R
\name{density_bins}
\alias{density_bins}
\alias{histogram_bins}
\title{Density bins and histogram bins as data frames}
\usage{
density_bins(x, n = 101, ...)

histogram_bins(x, n = 30, breaks = n, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Number of bins}

\item{...}{Additional arguments passed to \code{\link[=density]{density()}} or
\code{\link[=hist]{hist()}}.}

\item{breaks}{Used to set bins for \code{histogram_bins}. Can be number of bins (by default it is set to the value
of \code{n}) or a method for setting bins. See the \code{breaks} argument of \code{\link[=hist]{hist()}}.}
}
\value{
A data frame representing bins and their densities with the
following columns: \item{mid}{Bin midpoint} \item{lower}{Lower endpoint of
each bin} \item{upper}{Upper endpoint of each bin} \item{density}{Density
estimate of the bin}
}
\description{
Generates a data frame of bins representing the kernel density (or
histogram) of a vector, suitable for use in generating predictive
distributions for visualization. These functions were originally
designed for use with the now-deprecated \code{predict_curve()}, and
may be deprecated in the future.
}
\details{
These functions are simple wrappers to \code{\link[=density]{density()}} and
\code{\link[=hist]{hist()}} that compute density estimates and return their results
in a consistent format: a data frame of bins suitable for use with
the now-deprecated \code{\link[=predict_curve]{predict_curve()}}.

\code{density_bins} computes a kernel density estimate using
\code{\link[=density]{density()}}.

\code{histogram_bins} computes a density histogram using \code{\link[=hist]{hist()}}.
}
\seealso{
See \code{\link[=add_predicted_draws]{add_predicted_draws()}} and \code{\link[=stat_lineribbon]{stat_lineribbon()}} for a better approach. These
functions may be deprecated in the future.
}
\author{
Matthew Kay
}
\keyword{"Set2")}
\keyword{"drop")}
\keyword{%>%}
\keyword{))}
\keyword{*}
\keyword{+}
\keyword{-}
\keyword{.groups}
\keyword{1}
\keyword{1))}
\keyword{21,}
\keyword{=}
\keyword{\dontrun{}
\keyword{add_predicted_draws(m_mpg)}
\keyword{alpha}
\keyword{brm(mpg}
\keyword{by}
\keyword{c(0,}
\keyword{cyl,}
\keyword{data}
\keyword{data_grid(hp}
\keyword{density}
\keyword{fill}
\keyword{geom_point(aes(x}
\keyword{geom_rect(aes(}
\keyword{ggplot()}
\keyword{group_by(cyl)}
\keyword{hp}
\keyword{hp,}
\keyword{library(brms)}
\keyword{library(dplyr)}
\keyword{library(ggplot2)}
\keyword{library(modelr)}
\keyword{lower,}
\keyword{m_mpg}
\keyword{manip}
\keyword{mpg,}
\keyword{mtcars}
\keyword{mtcars)}
\keyword{ordered(cyl)),}
\keyword{ordered(cyl),}
\keyword{scale_alpha_continuous(range}
\keyword{scale_fill_brewer(palette}
\keyword{seq_range(hp,}
\keyword{shape}
\keyword{step}
\keyword{step))}
\keyword{step/2,}
\keyword{summarise(density_bins(.prediction),}
\keyword{theme_set(theme_light())}
\keyword{upper,}
\keyword{xmax}
\keyword{xmin}
\keyword{y}
\keyword{ymax}
\keyword{ymin}
\keyword{}}
\keyword{~}
