% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_draws.R
\name{sample_draws}
\alias{sample_draws}
\title{Sample draws from a tidy-format data frame of draws}
\usage{
sample_draws(data, ndraws, draw = ".draw", seed = NULL)
}
\arguments{
\item{data}{Data frame to sample from}

\item{ndraws}{The number of draws to return, or \code{NULL} to return all draws.}

\item{draw}{The name of the column indexing the draws; default \code{".draw"}.}

\item{seed}{A seed to use when subsampling draws (i.e. when \code{ndraws} is not \code{NULL}).}
}
\description{
Given a tidy-format data frame of draws with a column indexing each draw, subsample the data frame to a given size
based on a column indexing draws, ensuring that rows in sub-groups of a grouped data frame are sampled from the
same draws.
}
\details{
\code{sample_draws()} makes it easier to sub-sample a grouped, tidy-format data frame of draws. On a grouped data frame,
the naive approach of using \code{filter} with the \code{.draw} column will give incorrect results as it will
select a different sample within each group. \code{sample_draws()} ensures the same sample is selected within
each group.
}
\author{
Matthew Kay
}
\keyword{#}
\keyword{%>%}
\keyword{*}
\keyword{+}
\keyword{.draw)),}
\keyword{.epred,}
\keyword{/}
\keyword{0.25)}
\keyword{1}
\keyword{1,}
\keyword{100}
\keyword{101))}
\keyword{500)}
\keyword{=}
\keyword{NOTE:}
\keyword{\dontrun{}
\keyword{add_epred_draws(m_mpg)}
\keyword{all}
\keyword{alpha}
\keyword{always}
\keyword{and}
\keyword{best}
\keyword{brm(mpg}
\keyword{chain}
\keyword{chains}
\keyword{color}
\keyword{cyl,}
\keyword{data}
\keyword{data_grid(hp}
\keyword{do}
\keyword{draw}
\keyword{draws}
\keyword{example}
\keyword{few}
\keyword{fit}
\keyword{for}
\keyword{from}
\keyword{geom_line(aes(y}
\keyword{geom_point(data}
\keyword{ggplot(aes(x}
\keyword{group}
\keyword{group_by(cyl)}
\keyword{here}
\keyword{hp}
\keyword{hp,}
\keyword{in}
\keyword{intervals}
\keyword{is}
\keyword{it}
\keyword{iter}
\keyword{iterations}
\keyword{just}
\keyword{library(brms)}
\keyword{library(dplyr)}
\keyword{library(ggplot2)}
\keyword{library(modelr)}
\keyword{lines}
\keyword{m_mpg}
\keyword{making}
\keyword{manip}
\keyword{mpg,}
\keyword{mtcars}
\keyword{mtcars)}
\keyword{mtcars,}
\keyword{n}
\keyword{not}
\keyword{only}
\keyword{ordered(cyl)))}
\keyword{overplot}
\keyword{paste(cyl,}
\keyword{plots;}
\keyword{plotting}
\keyword{posterior}
\keyword{practice}
\keyword{quickly}
\keyword{runs}
\keyword{sample_draws}
\keyword{sample_draws(100)}
\keyword{seq_range(hp,}
\keyword{so}
\keyword{spaghetti}
\keyword{the}
\keyword{them}
\keyword{theme_set(theme_light())}
\keyword{to}
\keyword{use}
\keyword{when}
\keyword{y}
\keyword{}}
\keyword{~}
