% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdist-geom_slabinterval.R,
%   R/ggdist-geom_dotsinterval.R, R/ggdist-geom_interval.R,
%   R/ggdist-geom_lineribbon.R, R/ggdist-geom_pointinterval.R,
%   R/ggdist-lkjcorr_marginal.R, R/ggdist-parse_dist.R, R/ggdist-scales.R,
%   R/ggdist-stat_slabinterval.R, R/ggdist-stat_dist_slabinterval.R,
%   R/ggdist-stat_sample_slabinterval.R, R/ggdist-stat_dotsinterval.R,
%   R/ggdist-stat_pointinterval.R, R/ggdist-stat_interval.R,
%   R/ggdist-stat_lineribbon.R, R/ggdist-student_t.R, R/ggdist-theme_ggdist.R,
%   R/ggdist-tidy_format_translators.R, R/point_interval.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geom_slabinterval}
\alias{GeomSlabinterval}
\alias{geom_slab}
\alias{GeomSlab}
\alias{geom_dotsinterval}
\alias{GeomDotsinterval}
\alias{geom_dots}
\alias{GeomDots}
\alias{geom_interval}
\alias{GeomInterval}
\alias{geom_lineribbon}
\alias{GeomLineribbon}
\alias{geom_pointinterval}
\alias{GeomPointinterval}
\alias{dlkjcorr_marginal}
\alias{plkjcorr_marginal}
\alias{qlkjcorr_marginal}
\alias{rlkjcorr_marginal}
\alias{marginalize_lkjcorr}
\alias{parse_dist}
\alias{r_dist_name}
\alias{scale_point_colour_discrete}
\alias{scale_point_color_discrete}
\alias{scale_point_colour_continuous}
\alias{scale_point_color_continuous}
\alias{scale_point_fill_discrete}
\alias{scale_point_fill_continuous}
\alias{scale_point_alpha_continuous}
\alias{scale_point_alpha_discrete}
\alias{scale_point_size_continuous}
\alias{scale_point_size_discrete}
\alias{scale_interval_colour_discrete}
\alias{scale_interval_color_discrete}
\alias{scale_interval_colour_continuous}
\alias{scale_interval_color_continuous}
\alias{scale_interval_alpha_continuous}
\alias{scale_interval_alpha_discrete}
\alias{scale_interval_size_continuous}
\alias{scale_interval_size_discrete}
\alias{scale_interval_linetype_discrete}
\alias{scale_interval_linetype_continuous}
\alias{scale_slab_colour_discrete}
\alias{scale_slab_color_discrete}
\alias{scale_slab_colour_continuous}
\alias{scale_slab_color_continuous}
\alias{scale_slab_fill_discrete}
\alias{scale_slab_fill_continuous}
\alias{scale_slab_alpha_continuous}
\alias{scale_slab_alpha_discrete}
\alias{scale_slab_size_continuous}
\alias{scale_slab_size_discrete}
\alias{scale_slab_linetype_discrete}
\alias{scale_slab_linetype_continuous}
\alias{scale_slab_shape_discrete}
\alias{scale_slab_shape_continuous}
\alias{guide_colourbar2}
\alias{guide_colorbar2}
\alias{stat_slabinterval}
\alias{StatSlabinterval}
\alias{stat_dist_slabinterval}
\alias{StatDistSlabinterval}
\alias{stat_dist_halfeye}
\alias{stat_dist_eye}
\alias{stat_dist_ccdfinterval}
\alias{stat_dist_cdfinterval}
\alias{stat_dist_gradientinterval}
\alias{stat_dist_pointinterval}
\alias{stat_dist_interval}
\alias{stat_dist_slab}
\alias{stat_sample_slabinterval}
\alias{StatSampleSlabinterval}
\alias{stat_halfeye}
\alias{stat_eye}
\alias{stat_ccdfinterval}
\alias{stat_cdfinterval}
\alias{stat_gradientinterval}
\alias{stat_histinterval}
\alias{stat_slab}
\alias{stat_dotsinterval}
\alias{stat_dots}
\alias{stat_dist_dotsinterval}
\alias{stat_dist_dots}
\alias{stat_pointinterval}
\alias{stat_interval}
\alias{StatInterval}
\alias{stat_lineribbon}
\alias{stat_dist_lineribbon}
\alias{dstudent_t}
\alias{pstudent_t}
\alias{qstudent_t}
\alias{rstudent_t}
\alias{theme_ggdist}
\alias{theme_tidybayes}
\alias{facet_title_left_horizontal}
\alias{facet_title_right_horizontal}
\alias{axis_titles_bottom_left}
\alias{to_broom_names}
\alias{from_broom_names}
\alias{to_ggmcmc_names}
\alias{from_ggmcmc_names}
\alias{point_interval}
\alias{qi}
\alias{hdi}
\alias{Mode}
\alias{hdci}
\alias{mean_qi}
\alias{median_qi}
\alias{mode_qi}
\alias{mean_hdi}
\alias{median_hdi}
\alias{mode_hdi}
\alias{mean_hdci}
\alias{median_hdci}
\alias{mode_hdci}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggdist}{\code{\link[ggdist:theme_ggdist]{axis_titles_bottom_left}}, \code{\link[ggdist:lkjcorr_marginal]{dlkjcorr_marginal}}, \code{\link[ggdist:student_t]{dstudent_t}}, \code{\link[ggdist:theme_ggdist]{facet_title_left_horizontal}}, \code{\link[ggdist:theme_ggdist]{facet_title_right_horizontal}}, \code{\link[ggdist:tidy-format-translators]{from_broom_names}}, \code{\link[ggdist:tidy-format-translators]{from_ggmcmc_names}}, \code{\link[ggdist:geom_dotsinterval]{geom_dots}}, \code{\link[ggdist]{geom_dotsinterval}}, \code{\link[ggdist]{geom_interval}}, \code{\link[ggdist]{geom_lineribbon}}, \code{\link[ggdist]{geom_pointinterval}}, \code{\link[ggdist:geom_slabinterval]{geom_slab}}, \code{\link[ggdist]{geom_slabinterval}}, \code{\link[ggdist:ggdist-ggproto]{GeomDots}}, \code{\link[ggdist:ggdist-ggproto]{GeomDotsinterval}}, \code{\link[ggdist:ggdist-ggproto]{GeomInterval}}, \code{\link[ggdist:ggdist-ggproto]{GeomLineribbon}}, \code{\link[ggdist:ggdist-ggproto]{GeomPointinterval}}, \code{\link[ggdist:ggdist-ggproto]{GeomSlab}}, \code{\link[ggdist:ggdist-ggproto]{GeomSlabinterval}}, \code{\link[ggdist:scales]{guide_colorbar2}}, \code{\link[ggdist:scales]{guide_colourbar2}}, \code{\link[ggdist:point_interval]{hdci}}, \code{\link[ggdist:point_interval]{hdi}}, \code{\link[ggdist]{marginalize_lkjcorr}}, \code{\link[ggdist:point_interval]{mean_hdci}}, \code{\link[ggdist:point_interval]{mean_hdi}}, \code{\link[ggdist:point_interval]{mean_qi}}, \code{\link[ggdist:point_interval]{median_hdci}}, \code{\link[ggdist:point_interval]{median_hdi}}, \code{\link[ggdist:point_interval]{median_qi}}, \code{\link[ggdist:point_interval]{Mode}}, \code{\link[ggdist:point_interval]{mode_hdci}}, \code{\link[ggdist:point_interval]{mode_hdi}}, \code{\link[ggdist:point_interval]{mode_qi}}, \code{\link[ggdist]{parse_dist}}, \code{\link[ggdist:lkjcorr_marginal]{plkjcorr_marginal}}, \code{\link[ggdist]{point_interval}}, \code{\link[ggdist:student_t]{pstudent_t}}, \code{\link[ggdist:point_interval]{qi}}, \code{\link[ggdist:lkjcorr_marginal]{qlkjcorr_marginal}}, \code{\link[ggdist:student_t]{qstudent_t}}, \code{\link[ggdist:parse_dist]{r_dist_name}}, \code{\link[ggdist:lkjcorr_marginal]{rlkjcorr_marginal}}, \code{\link[ggdist:student_t]{rstudent_t}}, \code{\link[ggdist:scales]{scale_interval_alpha_continuous}}, \code{\link[ggdist:scales]{scale_interval_alpha_discrete}}, \code{\link[ggdist:scales]{scale_interval_color_continuous}}, \code{\link[ggdist:scales]{scale_interval_color_discrete}}, \code{\link[ggdist:scales]{scale_interval_colour_continuous}}, \code{\link[ggdist:scales]{scale_interval_colour_discrete}}, \code{\link[ggdist:scales]{scale_interval_linetype_continuous}}, \code{\link[ggdist:scales]{scale_interval_linetype_discrete}}, \code{\link[ggdist:scales]{scale_interval_size_continuous}}, \code{\link[ggdist:scales]{scale_interval_size_discrete}}, \code{\link[ggdist:scales]{scale_point_alpha_continuous}}, \code{\link[ggdist:scales]{scale_point_alpha_discrete}}, \code{\link[ggdist:scales]{scale_point_color_continuous}}, \code{\link[ggdist:scales]{scale_point_color_discrete}}, \code{\link[ggdist:scales]{scale_point_colour_continuous}}, \code{\link[ggdist:scales]{scale_point_colour_discrete}}, \code{\link[ggdist:scales]{scale_point_fill_continuous}}, \code{\link[ggdist:scales]{scale_point_fill_discrete}}, \code{\link[ggdist:scales]{scale_point_size_continuous}}, \code{\link[ggdist:scales]{scale_point_size_discrete}}, \code{\link[ggdist:scales]{scale_slab_alpha_continuous}}, \code{\link[ggdist:scales]{scale_slab_alpha_discrete}}, \code{\link[ggdist:scales]{scale_slab_color_continuous}}, \code{\link[ggdist:scales]{scale_slab_color_discrete}}, \code{\link[ggdist:scales]{scale_slab_colour_continuous}}, \code{\link[ggdist:scales]{scale_slab_colour_discrete}}, \code{\link[ggdist:scales]{scale_slab_fill_continuous}}, \code{\link[ggdist:scales]{scale_slab_fill_discrete}}, \code{\link[ggdist:scales]{scale_slab_linetype_continuous}}, \code{\link[ggdist:scales]{scale_slab_linetype_discrete}}, \code{\link[ggdist:scales]{scale_slab_shape_continuous}}, \code{\link[ggdist:scales]{scale_slab_shape_discrete}}, \code{\link[ggdist:scales]{scale_slab_size_continuous}}, \code{\link[ggdist:scales]{scale_slab_size_discrete}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_ccdfinterval}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_cdfinterval}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_ccdfinterval}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_cdfinterval}}, \code{\link[ggdist:geom_dotsinterval]{stat_dist_dots}}, \code{\link[ggdist:geom_dotsinterval]{stat_dist_dotsinterval}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_eye}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_gradientinterval}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_halfeye}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_interval}}, \code{\link[ggdist:stat_lineribbon]{stat_dist_lineribbon}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_pointinterval}}, \code{\link[ggdist:stat_dist_slabinterval]{stat_dist_slab}}, \code{\link[ggdist]{stat_dist_slabinterval}}, \code{\link[ggdist:geom_dotsinterval]{stat_dots}}, \code{\link[ggdist:geom_dotsinterval]{stat_dotsinterval}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_eye}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_gradientinterval}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_halfeye}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_histinterval}}, \code{\link[ggdist]{stat_interval}}, \code{\link[ggdist]{stat_lineribbon}}, \code{\link[ggdist]{stat_pointinterval}}, \code{\link[ggdist]{stat_sample_slabinterval}}, \code{\link[ggdist:stat_sample_slabinterval]{stat_slab}}, \code{\link[ggdist]{stat_slabinterval}}, \code{\link[ggdist:ggdist-ggproto]{StatDistSlabinterval}}, \code{\link[ggdist:ggdist-ggproto]{StatInterval}}, \code{\link[ggdist:ggdist-ggproto]{StatSampleSlabinterval}}, \code{\link[ggdist:ggdist-ggproto]{StatSlabinterval}}, \code{\link[ggdist]{theme_ggdist}}, \code{\link[ggdist:theme_ggdist]{theme_tidybayes}}, \code{\link[ggdist:tidy-format-translators]{to_broom_names}}, \code{\link[ggdist:tidy-format-translators]{to_ggmcmc_names}}}
}}

